fix: 코드 리팩터링
fix
- fetch 사용 서비스 함수 2개 service파일로 분리
